requirements.txt
torch==2.2.0
numpy==1.26.4
gymnasium==0.29.1
ray[rllib]==2.9.0
pyyaml==6.0.1
matplotlib==3.8.3
pandas==2.2.1
tqdm==4.66.2
